requirements.txt
pyzmq
opencv-python
pillow